requirements.txt
tensorflow==2.18.0
pandas
gdown
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box==7.3.0
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .